clarify component names; progress grid layout
