+++ src/module/vue/components/bonds.vue
@@ -5,14 +5,14 @@
       <btn-faicon class="block nogrow" icon="edit" @click="editBonds" />
       <btn-faicon class="block nogrow" icon="dice-d6" @click="rollBonds" />
     </div>
-    <progress-track :ticks="bondcount" />
+    <Track :ticks="bondcount" />
   </div>
 </template>
 
 <script setup lang="ts">
 import { inject, computed, Ref } from 'vue'
 import { $ActorKey } from '../provisions'
-import ProgressTrack from './progress/progress-track.vue'
+import Track from './progress/track.vue'
 import btnFaicon from './buttons/btn-faicon.vue'
 
 const actor = inject('actor') as Ref
