+++ src/module/vue/components/character-sheet-tabs/ironsworn-main.vue
@@ -63,7 +63,7 @@ import Bonds from '../bonds.vue'
 import OrderButtons from '../order-buttons.vue'
 import Asset from '../asset/asset.vue'
 import BtnCompendium from '../buttons/btn-compendium.vue'
-import ProgressBox from '../progress/progress-box.vue'
+import ProgressListItem from '../progress/progress-list-item.vue'
 import ProgressControls from '../progress-controls.vue'
 import { throttle } from 'lodash'
 import BtnFaicon from '../buttons/btn-faicon.vue'
