+++ src/module/vue/components/character-sheet-tabs/sf-connections.vue
@@ -13,7 +13,7 @@
           @sortUp="sortUp"
           @sortDown="sortDown"
         />
-        <progress-box :item="item" :show-star="true" />
+        <progress-list-item :item="item" :show-star="true" />
       </div>
     </transition-group>
 
@@ -36,7 +36,7 @@
 import { computed, inject, Ref } from 'vue'
 import { $ActorKey } from '../../provisions'
 import OrderButtons from '../order-buttons.vue'
-import ProgressBox from '../progress/progress-box.vue'
+import ProgressListItem from '../progress/progress-list-item.vue'
 import BtnFaicon from '../buttons/btn-faicon.vue'
 import { ProgressDataProperties } from '../../../item/itemtypes'
