+++ src/module/vue/components/character-sheet-tabs/sf-legacies.vue
@@ -5,7 +5,7 @@
     <legacy-track propKey="discoveries" :title="$t('IRONSWORN.Discoveries')" />
 
     <hr class="nogrow" v-if="starredProgresses.length > 0" />
-    <progress-box
+    <progress-list-item
       v-for="item in starredProgresses"
       :key="item._id"
       :item="item"
@@ -16,7 +16,7 @@
 <script lang="ts" setup>
 import { computed, inject, Ref } from 'vue'
 import legacyTrack from '../xp/legacy-track.vue'
-import progressBox from '../progress/progress-box.vue'
+import ProgressListItem from '../progress/progress-list-item.vue'
 
 const actor = inject('actor') as Ref
