+++ src/module/vue/components/progress/track-box.vue
@@ -1,29 +1,47 @@
 <template>
-  <svg
-    class="track-box-marks"
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 100 100"
+  <div
+    class="flexcol track-box"
     :aria-valuenow="ticks"
+    :aria-valuetext="`${ticks} ticks`"
   >
-    <TransitionGroup name="draw-tick">
-      <line
-        v-bind="tickProps"
-        v-for="(tick, i) in tickRange"
-        :key="`tick-${tick}`"
-        v-show="props.ticks > i"
-        :transform="tickTransforms[i]"
-        :class="`tick-${tick}`"
-        :data-tick="tick"
-      />
-    </TransitionGroup>
-  </svg>
+    <svg
+      class="track-box-marks"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 100 100"
+      role="presentational"
+    >
+      <TransitionGroup name="draw-tick">
+        <line
+          v-bind="tickProps"
+          v-for="(tick, i) in tickRange"
+          :key="`tick-${tick}`"
+          v-show="props.ticks > i"
+          :transform="tickTransforms[i]"
+          :class="`tick-${tick}`"
+          :data-tick="tick"
+        />
+      </TransitionGroup>
+    </svg>
+  </div>
 </template>
 <style lang="less">
+.track-box {
+  border: 1px solid;
+  align-items: center;
+  justify-content: center;
+  aspect-ratio: 1;
+  border-radius: 3px;
+  object-fit: contain;
+  box-shadow: 2px 2px 0 currentColor;
+  .track-box-marks {
+    margin: 5px;
+  }
+}
 .track-box-marks {
   aspect-ratio: 1;
   stroke: currentColor;
-  stroke-width: 2px;
   overflow: visible;
+  stroke-width: 2px;
 }
 
 .progress-tick {
@@ -33,6 +51,7 @@
   stroke-dasharray: 50%;
 }
 
+// Progress tick draw animation
 .draw-tick-enter-active {
   transition: 0.8s stroke-dashoffset, stroke-dasharray;
 }
@@ -53,7 +72,7 @@
 .draw-tick-enter-to {
   stroke-dashoffset: 0;
 }
-
+// helper mixin functions
 .animateTick(@value,@duration,@delay:0s) {
   .draw-tick-enter-active[data-tick='@{value}'] {
     transition-duration: @duration;
@@ -69,7 +88,8 @@
 }
 .track {
   &[data-rank='1'] {
-    // troublesome: marks 3 boxes (12 ticks)
+    // see the Track component
+    // Challenge rank troublesome: marks 3 boxes (12 ticks)
     @box1: 1s;
     @box2: 1s;
     @box3: 1s;
@@ -83,8 +103,10 @@
       @d1: (@box1+ @box2);
       .animateBox(@box3,@d1);
     }
+    // handles edge cases when
   }
   &[data-rank='2'] {
+    // Challenge rank dangerous: marks 2 boxes (8 ticks)
     @box1: 1s;
     @box2: 0.75s;
     .track-box:nth-child(2n + 1) {
@@ -95,11 +117,11 @@
     }
   }
   &[data-rank='3'] {
+    // Challenge rank formidable: marks 1 box (4 ticks).
     .animateBox(1s);
-    // formidable: marks 4 ticks at a time
   }
   &[data-rank='4'] {
-    // extreme: marks 2 ticks at a time.
+    // Challenge rank extreme: marks 2 ticks.
     @d1: 0.5s;
     @d2: 0.5s;
     .animateTick(1,@d1);
@@ -108,7 +130,7 @@
     .animateTick(4,@d2,@d1);
   }
   &[data-rank='5'] {
-    // epic: marks 1 tick at a time, so doesn't need anything fancy.
+    // Challenge rank epic: marks 1 tick.
     .draw-tick-enter-active {
       transition-duration: 0.5s;
     }
